requirements.txt
django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
faiss-cpu==1.11.0.post1
numpy==2.3.1
requests==2.32.4
sentence-transformers==5.0.0